+++ docassemble_webapp/docassemble/webapp/server.py
@@ -7,7 +7,9 @@ import email as emailpackage
 import filecmp
 import fnmatch
 import hashlib
+from pathlib import Path
 import importlib
+import importlib.resources
 import inspect
 from io import BytesIO, TextIOWrapper
 import json
@@ -117,7 +119,6 @@ import links_from_header
 import oauth2client.client
 import pandas
 from PIL import Image
-import pkg_resources
 import pyotp
 from pygments import highlight
 from pygments.formatters import HtmlFormatter  # pylint: disable=no-name-in-module
@@ -602,10 +603,8 @@ PAGINATION_LIMIT_PLUS_ONE = PAGINATION_LIMIT + 1
 
 # PLAYGROUND_MODULES_DIRECTORY = daconfig.get('playground_modules', )
 
-init_py_file = """try:
-    __import__('pkg_resources').declare_namespace(__name__)
-except ImportError:
-    __path__ = __import__('pkgutil').extend_path(__path__, __name__)
+init_py_file = """
+__import__('pkg_resources').declare_namespace(__name__)
 """
 
 # if not os.path.isfile(os.path.join(PLAYGROUND_MODULES_DIRECTORY, 'docassemble', '__init__.py')):
@@ -16925,10 +16924,7 @@ def create_package():
     if request.method == 'POST' and form.validate():
         pkgname = re.sub(r'^docassemble-', r'', form.name.data)
         initpy = """\
-try:
-    __import__('pkg_resources').declare_namespace(__name__)
-except ImportError:
-    __path__ = __import__('pkgutil').extend_path(__path__, __name__)
+__import__('pkg_resources').declare_namespace(__name__)
 
 """
         licensetext = """\
@@ -22465,7 +22461,7 @@ def server_error(the_error):
 
 @app.route('/bundle.css', methods=['GET'])
 def css_bundle():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['bootstrap-fileinput', 'css', 'fileinput.min.css'], ['labelauty', 'source', 'jquery-labelauty.min.css'], ['bootstrap-combobox', 'css', 'bootstrap-combobox.min.css'], ['bootstrap-slider', 'dist', 'css', 'bootstrap-slider.min.css'], ['app', 'app.min.css']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22476,7 +22472,7 @@ def css_bundle():
 
 @app.route('/playgroundbundle.css', methods=['GET'])
 def playground_css_bundle():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['codemirror', 'lib', 'codemirror.css'], ['codemirror', 'addon', 'search', 'matchesonscrollbar.css'], ['codemirror', 'addon', 'display', 'fullscreen.css'], ['codemirror', 'addon', 'scroll', 'simplescrollbars.css'], ['codemirror', 'addon', 'hint', 'show-hint.css'], ['app', 'pygments.min.css'], ['bootstrap-fileinput', 'css', 'fileinput.min.css']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22487,7 +22483,7 @@ def playground_css_bundle():
 
 @app.route('/bundle.js', methods=['GET'])
 def js_bundle():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['app', 'jquery.min.js'], ['app', 'jquery.validate.min.js'], ['app', 'additional-methods.min.js'], ['app', 'jquery.visible.min.js'], ['bootstrap', 'js', 'bootstrap.bundle.min.js'], ['bootstrap-slider', 'dist', 'bootstrap-slider.min.js'], ['labelauty', 'source', 'jquery-labelauty.min.js'], ['bootstrap-fileinput', 'js', 'plugins', 'piexif.min.js'], ['bootstrap-fileinput', 'js', 'fileinput.min.js'], ['bootstrap-fileinput', 'themes', 'fas', 'theme.min.js'], ['app', 'app.min.js'], ['bootstrap-combobox', 'js', 'bootstrap-combobox.min.js'], ['app', 'socket.io.min.js']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22498,7 +22494,7 @@ def js_bundle():
 
 @app.route('/playgroundbundle.js', methods=['GET'])
 def playground_js_bundle():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['areyousure', 'jquery.are-you-sure.js'], ['codemirror', 'lib', 'codemirror.js'], ['codemirror', 'addon', 'search', 'searchcursor.js'], ['codemirror', 'addon', 'scroll', 'annotatescrollbar.js'], ['codemirror', 'addon', 'search', 'matchesonscrollbar.js'], ['codemirror', 'addon', 'display', 'fullscreen.js'], ['codemirror', 'addon', 'edit', 'matchbrackets.js'], ['codemirror', 'addon', 'hint', 'show-hint.js'], ['codemirror', 'mode', 'yaml', 'yaml.js'], ['codemirror', 'mode', 'python', 'python.js'], ['yamlmixed', 'yamlmixed.js'], ['codemirror', 'mode', 'markdown', 'markdown.js'], ['bootstrap-fileinput', 'js', 'plugins', 'piexif.min.js'], ['bootstrap-fileinput', 'js', 'fileinput.min.js'], ['bootstrap-fileinput', 'themes', 'fas', 'theme.min.js']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22509,7 +22505,7 @@ def playground_js_bundle():
 
 @app.route('/adminbundle.js', methods=['GET'])
 def js_admin_bundle():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['app', 'jquery.min.js'], ['bootstrap', 'js', 'bootstrap.bundle.min.js']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22520,7 +22516,7 @@ def js_admin_bundle():
 
 @app.route('/bundlewrapjquery.js', methods=['GET'])
 def js_bundle_wrap():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = '(function($) {'
     for parts in [['app', 'jquery.validate.min.js'], ['app', 'additional-methods.min.js'], ['app', 'jquery.visible.js'], ['bootstrap', 'js', 'bootstrap.bundle.min.js'], ['bootstrap-slider', 'dist', 'bootstrap-slider.min.js'], ['bootstrap-fileinput', 'js', 'plugins', 'piexif.min.js'], ['bootstrap-fileinput', 'js', 'fileinput.min.js'], ['bootstrap-fileinput', 'themes', 'fas', 'theme.min.js'], ['app', 'app.min.js'], ['labelauty', 'source', 'jquery-labelauty.min.js'], ['bootstrap-combobox', 'js', 'bootstrap-combobox.min.js'], ['app', 'socket.io.min.js']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp:
@@ -22532,7 +22528,7 @@ def js_bundle_wrap():
 
 @app.route('/bundlenojquery.js', methods=['GET'])
 def js_bundle_no_query():
-    base_path = pkg_resources.resource_filename(pkg_resources.Requirement.parse('docassemble.webapp'), os.path.join('docassemble', 'webapp', 'static'))
+    base_path = Path(importlib.resources.files('docassemble.webapp'), 'data', 'static')
     output = ''
     for parts in [['app', 'jquery.validate.min.js'], ['app', 'additional-methods.min.js'], ['app', 'jquery.visible.min.js'], ['bootstrap', 'js', 'bootstrap.bundle.min.js'], ['bootstrap-slider', 'dist', 'bootstrap-slider.min.js'], ['bootstrap-fileinput', 'js', 'plugins', 'piexif.min.js'], ['bootstrap-fileinput', 'js', 'fileinput.min.js'], ['bootstrap-fileinput', 'themes', 'fas', 'theme.min.js'], ['app', 'app.min.js'], ['labelauty', 'source', 'jquery-labelauty.min.js'], ['bootstrap-combobox', 'js', 'bootstrap-combobox.min.js'], ['app', 'socket.io.min.js']]:
         with open(os.path.join(base_path, *parts), encoding='utf-8') as fp: